requirements.txt
streamlit
opencv-python
ultralytics
numpy
pillow
torch